+++ internal/metrics/batching/metrics_test.go
