+++ internal/metrics/batching/testdata/batching_happy.txt
