+++ internal/metrics/batching/testdata/batching_nometrics.txt
