+++ internal/metrics/readers/testdata/readers_happy.txt
