+++ internal/metrics/readers/testdata/readers_nometrics.txt
